+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Add input component - [#40](https://github.com/PrefectHQ/miter-design/pull/40)
 - Enforce slot styling from card component - [#57](https://github.com/PrefectHQ/miter-design/pull/57)
 - Add drop shadow options to card - [#53](https://github.com/PrefectHQ/miter-design/pull/53)
 - Don't hard code card padding - [#56](https://github.com/PrefectHQ/miter-design/pull/56)